Changes the behavior of the cursor and the ruler.
When the cursor blinking setting is set so that the cursor does not blink,
if the error indicator is activated, the cursor disappears from the screen.

At the same time, when the ruler is activated, if the cursor blinks, the ruler also blinks.

This patch changes both of these situations.